% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/exact_match.R
\name{exact_match}
\alias{exact_match}
\title{Exact matching to WCVP.}
\usage{
exact_match(names_df, wcvp_names, name_col, author_col = NULL, id_col)
}
\arguments{
\item{names_df}{a data frame of names for matching.}

\item{wcvp_names}{a data frame of WCVP names to match against,
can be loaded from \link[rWCVPdata:wcvp_names]{rWCVPdata::wcvp_names}.}

\item{name_col}{the column in \code{names_df} that has the taxon name for matching.}

\item{author_col}{the column in \code{names_df} that has the name authority, to aid
matching. Set to \code{NULL} to match with no author string.}

\item{id_col}{the column in \code{names_df} that has the observation id.}
}
\value{
Match results from WCVP bound to the original data from \code{names_df}.
}
\description{
Exact matching of names to the WCVP, optionally using the author string to
refine results.
}
\examples{
wcvp_names <- rWCVPdata::wcvp_names

# including author string
exact_match(redlist_example, wcvp_names, "scientificName", author_col="authority",
id_col = "assessmentId")

# without author string
exact_match(redlist_example, wcvp_names, "scientificName", id_col = "assessmentId")

}
\seealso{
Other name matching functions: 
\code{\link{fuzzy_match}()},
\code{\link{match_names}()}
}
\concept{name matching functions}
