% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/wcvp_fuzzy_match.R
\name{fuzzy_match}
\alias{fuzzy_match}
\title{Distance-based matching using Levenshtein similarity DEEXPORT WHEN FINISHED}
\usage{
fuzzy_match(x, wcvp_to_search)
}
\arguments{
\item{x}{data.frame containing the variable \code{taxon.name} to match}

\item{wcvp_to_search}{data.frame containing selected variables from the WVCP,
without genus-level names}
}
\value{
data.frame with relevant WCVP information, plus match_type - an
informative column giving the match result and Levenshtein similarity (as
implemented in the \code{RecordLinkage} package)
}
\description{
Distance-based matching using Levenshtein similarity DEEXPORT WHEN FINISHED
}
