% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/generate_occurrence_matrix.r
\name{generate_occurrence_matrix}
\alias{generate_occurrence_matrix}
\title{Generate occurrence matrix for taxa and areas}
\usage{
generate_occurrence_matrix(
  taxon = NULL,
  taxon.rank = c("species", "genus", "family"),
  area = NULL,
  native = TRUE,
  introduced = TRUE,
  extinct = TRUE,
  location_doubtful = TRUE,
  local_wcvp = FALSE,
  wcvp_names = NULL,
  wcvp_distributions = NULL
)
}
\arguments{
\item{taxon}{Character. One or many taxa to be included. Defaults to NULL (all species)}

\item{taxon.rank}{Character. One of "species", "genus" or "family" giving the rank of the value/s in \code{taxon}. Must be specified unless taxon is NULL. Note that this must be one value only - it is not possible to mix and match families and genera, for example.}

\item{area}{Character. One or many WGSPRD level 3 region codes. Defaults to NULL (global).}

\item{native}{Logical. Include species occurrences not flagged as introduced, extinct or doubtful? Defaults to TRUE.}

\item{introduced}{Logical. Include species occurrences flagged as introduced? Defaults to TRUE.}

\item{extinct}{Logical. Include species occurrences flagged as extinct? Defaults to TRUE.}

\item{location_doubtful}{Logical. Include species occurrences flagged as location doubtful? Defaults to TRUE.}

\item{local_wcvp}{Logical. If FALSE (the default), use data from \code{rWCVPdata}.
If TRUE, use a local copy of the data (useful if rWCVPdata is not the latest
version of the checklist).}

\item{wcvp_names}{Pointer to the WCVP names dataset. Ignored if \code{local.wcvp = FALSE}. Defaults to NULL.}

\item{wcvp_distributions}{Pointer to the WCVP distributions dataset. Ignored if \code{local.wcvp = FALSE}. Defaults to NULL.}
}
\value{
A data.frame containing the \code{taxon_name} and \code{plant_name_id}
of all species that are present in the \code{area}, plus one variable for each WGSPRD level 3 region in \code{area}, with species presences marked as 1 and absences marked as 0.
}
\description{
Generate occurrence matrix for taxa and areas
}
\details{
See vignette "Generating occurrence matrices with rWCVP" for an example of how this output can be formatted for publication.
}
\examples{
generate_occurrence_matrix(taxon="Poa",taxon.rank="genus",
area=c("TAS", "VIC","NSW"), introduced=FALSE)


}
