% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/wgsrpd3_code_mapping.R
\name{get_area_name}
\alias{get_area_name}
\title{Get area description from vector of area codes}
\usage{
get_area_name(area.codes)
}
\arguments{
\item{area.codes}{Character vector containing the set of codes to be mapped to a name.}
}
\value{
Character. Either a vector of length one, with a name for the set of
Level 3 areas, or (if no name exists for that set of areas) the input vector of codes.
}
\description{
Get area description from vector of area codes
}
\examples{
get_area_name(get_wgsrpd3_codes("Brazil"))
}
