% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/get_distribution.R
\name{get_distribution}
\alias{get_distribution}
\title{Generate spatial distribution objects for species, genera or families}
\usage{
get_distribution(
  taxon,
  rank = c("species", "genus", "family"),
  native = TRUE,
  introduced = TRUE,
  extinct = TRUE,
  location_doubtful = TRUE,
  local_wcvp = FALSE,
  wcvp_names = NULL,
  wcvp_distributions = NULL
)
}
\arguments{
\item{taxon}{Character. The taxon to be mapped.}

\item{rank}{Character. The rank of \code{taxon}. Note: if multiple taxa are
to be mapped, they must be of the same rank.}

\item{native}{Logical. Include native range? Defaults to TRUE.}

\item{introduced}{Logical. Include introduced range? Defaults to TRUE.}

\item{extinct}{Logical. Include extinct range? Defaults to TRUE.}

\item{location_doubtful}{Logical. Include occurrences that are thought to be
doubtful? Defaults to TRUE.}

\item{local_wcvp}{Logical. If FALSE (the default), use data from \code{rWCVPdata}.
If TRUE, use a local copy of the data (useful if rWCVPdata is not the latest
version of the checklist).}

\item{wcvp_names}{Pointer to the WCVP names dataset. Ignored if \code{local.wcvp = FALSE}. Defaults to NULL.}

\item{wcvp_distributions}{Pointer to the WCVP distributions dataset. Ignored if \code{local.wcvp = FALSE}. Defaults to NULL.}
}
\value{
A \code{sf} data.frame containing the range polygon/s of the taxon.
}
\description{
Generate spatial distribution objects for species, genera or families
}
\details{
For an exmaple of how this function can be used to clean occurrence records using a spatial filter, see the vignette "Cleaning occurrence records with rWCVP".
}
\examples{
r <- get_distribution("Callitris", rank="genus")
p <- plot_distribution(r)
p
}
