% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/wcvp_fuzzy_match.R
\name{phonetic_match}
\alias{phonetic_match}
\title{Phonetic matching - DEEXPORT WHEN FINISHED}
\usage{
phonetic_match(x, wcvp_to_search)
}
\arguments{
\item{x}{data.frame containing the variable \code{taxon.name} to match}

\item{wcvp_to_search}{data.frame containing selected variables from the WVCP,
without genus-level names}
}
\value{
data.frame with relevant WCVP information, plus match_type - an
informative column giving the match result and Levenshtein similarity to
original name (note that this is not the metric that is used to link the
names, see \code{fuzzy_match} for Levenshtein matching)
}
\description{
Phonetic matching - DEEXPORT WHEN FINISHED
}
