% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plot_distribution.R
\name{plot_distribution}
\alias{plot_distribution}
\title{Plot distribution maps for species, genera or families}
\usage{
plot_distribution(
  range,
  native = TRUE,
  introduced = TRUE,
  extinct = TRUE,
  location_doubtful = TRUE
)
}
\arguments{
\item{range}{sf object of the type output by \code{get_distribution()}.}

\item{native}{Logical. Include native range? Defaults to TRUE.}

\item{introduced}{Logical. Include introduced range? Defaults to TRUE.}

\item{extinct}{Logical. Include extinct range? Defaults to TRUE.}

\item{location_doubtful}{Logical. Include occurrences that are thought to be
doubtful? Defaults to TRUE.}
}
\value{
List of length two, with global and local maps of distribution as
ggplot objects. If taxon has global distribution, only the first element is returned.
}
\description{
Plot distribution maps for species, genera or families
}
\details{
The colour scheme mirrors that used by Plants of the World (POWO;
https://powo.science.kew.org/).
}
\examples{
p <- plot_distribution(get_distribution("Callitris", rank="genus"))
p[[1]]
p[[2]]
#now only the native range
p <- plot_distribution(get_distribution("Callitris", rank="genus"), introduced=FALSE)
p[[1]]
p[[2]]
}
