% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plot_distribution.R
\name{powo_map}
\alias{powo_map}
\title{Plot a POWO style map for given range and range centroids.}
\usage{
powo_map(range_sf, centroids_sf)
}
\arguments{
\item{range_sf}{A simple features data frame of range polygons}

\item{centroids_sf}{A simple features data frame of range centroids}
}
\value{
A ggplot map of the range
}
\description{
Plot a POWO style map for given range and range centroids.
}
