% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{redlist_example}
\alias{redlist_example}
\title{Example dataset for name matching}
\format{
A data frame with 20 rows and 4 variables:
\describe{
  \item{assessmentId}{Red List identifier}
  \item{scientificName}{Taxon name.}
  \item{redlistCategory}{Red List threat category}
  \item{authority}{Taxon author/s.}
  }
}
\source{
Downloaded and sampled from \url{https://www.iucnredlist.org/}
}
\usage{
redlist_example
}
\description{
A dataset containing 20 sampled Red List assessments for name-matching
}
\keyword{datasets}
