% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/summary_table.r
\name{summary_table}
\alias{summary_table}
\title{Generate a summary table from the WCVP}
\usage{
summary_table(
  taxon = NULL,
  rank = c("species", "genus", "family", "order", "higher"),
  area = NULL,
  return = c("data.frame", "gt_table"),
  grouping.var = c("area_code_l3", "genus", "family", "order", "higher"),
  wcvp_names = NULL,
  wcvp_distributions = NULL
)
}
\arguments{
\item{taxon}{Character. Taxon to be included. Defaults to NULL (no taxonomic filter; all taxa).}

\item{rank}{Character. One of "species", "genus", "family", "order" or "higher", giving the rank of the value/s in \code{taxon}. Must be specified unless taxon is \code{NULL}.}

\item{area}{Character. One or many WGSPRD level 3 region codes. Defaults to \code{NULL} (global).}

\item{return}{Return the checklist as a \code{gt} table \code{"gt_table"} or \code{"data.frame"}? Defaults to \code{"data.frame"}.}

\item{grouping.var}{Character; one of \code{"area_code_l3", "genus", "family","order"} or \code{"higher"} specifying how the summary should be arranged. Defaults to \code{area_code_l3}.}

\item{wcvp_names}{Pointer to the WCVP names dataset. Defaults to \code{NULL} (uses data from \code{WCVPdata}).}

\item{wcvp_distributions}{Pointer to the WCVP distributions dataset. Defaults to \code{NULL} (uses data from \code{WCVPdata}).}
}
\value{
Data.frame with filtered data, or a \code{gt} table
}
\description{
Generate a summary table from the WCVP
}
\examples{
ferns <- summary_table("Ferns", "higher", get_wgsrpd3_codes("New Zealand"), grouping.var="family")
summary_table_gt(ferns)

}
