% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/summary_table.r
\name{summary_table_gt}
\alias{summary_table_gt}
\title{Render a summary table from \link{summary_table}}
\usage{
summary_table_gt(x)
}
\arguments{
\item{x}{List.}
}
\value{
\link{gt} table
}
\description{
Render a summary table from \link{summary_table}
}
\examples{
ferns <- summary_table("Ferns", "higher", get_wgsrpd3_codes("New Zealand"), grouping.var="family")
summary_table_gt(ferns)

}
