% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{taxonomic_mapping}
\alias{taxonomic_mapping}
\title{Data for mapping plant family to order or higher classification}
\format{
A data frame with 457 rows and 3 variables: \code{family, order} and \code{higher_classification}
}
\source{
Fern and lycophyte taxonomy from PPG I (2016; \doi{10.1111/jse.12229}). Angiosperm taxonomy from APG IV (2016; \doi{10.1111/boj.12385}). Gymnosperm taxonomy from Forest et al (2018; \doi{10.1038/s41598-018-24365-4})
}
\usage{
taxonomic_mapping
}
\description{
A dataset containing the higher classification (Angiosperms, Gymnosperms, Ferns and Lycophytes) and Order for each family in the WCVP.
}
\keyword{datasets}
