% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/wcvp_fuzzy_match.R
\name{wcvp_fuzzy_match}
\alias{wcvp_fuzzy_match}
\title{Fuzzy (approximate) matching to the WCVP}
\usage{
wcvp_fuzzy_match(x)
}
\arguments{
\item{x}{data.frame containing the variables \code{id, taxon.name} to match}
}
\value{
data.frame consisting of \code{x} bound to \code{plant_name_id, taxon_name,
taxon_rank, taxon_status,}
\code{accepted_plant_name_id, homotypic_synonym} and
\code{match_type} (a descriptor of the match result)
}
\description{
Fuzzy (approximate) matching to the WCVP
}
\details{
This function first uses phonetic matching, using metaphone encoding
(via the \code{phonics} package) with a maximum codelength of 20. For names
that remain unmatched, fuzzy matching is performed using the Levenshtein
edit distance and similarity (standardised edit distance). It is strongly recommended that this function only be called using \code{wcvp_name_match}, as it is much more efficient to find any exact matches before proceeding to fuzzy matching.
}
\examples{
#column names must include id and taxon.name
colnames(redlist_example)[1:2] <- c("id","taxon.name")
wcvp_fuzzy_match(redlist_example)

}
