% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/wcvp_name_match.R
\name{wcvp_name_match}
\alias{wcvp_name_match}
\title{Match names to the WCVP}
\usage{
wcvp_name_match(
  x,
  taxon.name = NULL,
  id = NULL,
  authors = NULL,
  genus = NULL,
  spec = NULL,
  infra.rank = NULL,
  infra = NULL
)
}
\arguments{
\item{x}{data.frame of names to match}

\item{taxon.name}{Character. The name of the variable in \code{x}
containing the genus/binomial/trinomial taxon name. Defaults to NULL.
Note: if using default, \code{genus} and \code{spec} \emph{must} be
supplied.}

\item{id}{Character. The name of the variable in \code{x} containing the
identifier (e.g. Red List assessment ID). Defaults to NULL.}

\item{authors}{Character. The name of the variable in \code{x} containing
the author information. Defaults to NULL.}

\item{genus}{Character. The name of the variable in \code{x} containing the
genus name. Defaults to NULL.}

\item{spec}{Character. The name of the variable in \code{x} containing the
specific epithet. Defaults to NULL.}

\item{infra.rank}{Character. The name of the variable in \code{x} containing the
infraspecific rank (e.g. subsp., var.). Defaults to NULL.}

\item{infra}{Character. The name of the variable in \code{x} containing the
infraspecific epithet. Defaults to NULL.}
}
\value{
data.frame consisting of the input \code{x} with name matching variables appended
}
\description{
Match names to the WCVP
}
\examples{
# without author
wcvp_name_match(redlist_example, taxon.name = "scientificName")
# with author
wcvp_name_match(redlist_example, taxon.name = "scientificName", authors = "authority")
}
