% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/wcvp_reformat.r
\name{wcvp_reformat}
\alias{wcvp_reformat}
\title{Reformat local versions of WCVP}
\usage{
wcvp_reformat(wcvp_local, version = NULL)
}
\arguments{
\item{wcvp_local}{Data.frame. Local copy of the WCVP.}

\item{version}{Either 9 or "v9". We will add support for other versions as needed.}
}
\value{
A data.frame with the same variable structure as the WCVP that is
included in the data package \code{rWCVPdata}.
}
\description{
Reformat local versions of WCVP
}
\details{
Note that not all of the original variables are preserved during reformatting.
For example, publication is a single variable in v9, but split over multiple in
the data package. It is therefore not possible to simply rename this variable.
Variables that are present in the data package but not in v9 are filled with \code{NA}.
}
