% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/wcvp_summary.r
\name{wcvp_summary_gt}
\alias{wcvp_summary_gt}
\title{Render a summary table from \link{wcvp_summary}}
\usage{
wcvp_summary_gt(x)
}
\arguments{
\item{x}{List.}
}
\value{
\link{gt} table
}
\description{
Render a summary table from \link{wcvp_summary}
}
\examples{
\dontrun{
ferns <- wcvp_summary("Ferns", "higher", get_wgsrpd3_codes("New Zealand"), grouping_var = "family")
wcvp_summary_gt(ferns)
}
}
