% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{wgsrpd3}
\alias{wgsrpd3}
\title{Biodiversity Information Standards (TDWG) World Geographical Scheme for
Recording Plant Distributions (WGSRPD)}
\format{
An 'sf' object with 20 rows and 4 variables:
\describe{
\item{LEVEL3_NAM}{Region name}
\item{LEVEL3_COD}{Region code}
\item{LEVEL2_COD}{Level 2 code}
\item{LEVEL1_COD}{Level 1 code (continent)}
\item{geometry}{sf geometry}
\item{fillcol}{Used for mapping.}
}
}
\source{
\url{https://github.com/tdwg/wgsrpd/tree/master/level3}
}
\usage{
wgsrpd3
}
\description{
Spatial data for WGSRPD Level 3, for plotting maps
}
\keyword{datasets}
