% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/fuzzy_match.R
\name{fuzzy_match}
\alias{fuzzy_match}
\alias{phonetic_match}
\alias{edit_match}
\title{Fuzzy (approximate) matching to the WCVP.}
\usage{
fuzzy_match(names_df, wcvp, name_col)

phonetic_match(names_df, wcvp, name_col)

edit_match(names_df, wcvp, name_col)
}
\arguments{
\item{names_df}{a data frame of names for matching.}

\item{wcvp}{a data frame of WCVP names to match against.}

\item{name_col}{the column in \code{names_df} that has the taxon name for matching.}
}
\value{
Match results from WCVP bound to the original data from \code{names_df}.
}
\description{
Fuzzy matching to names in the WCVP using phonetic matching and edit distance.
The WCVP can be loaded for matching from \link[rWCVPdata:wcvp_names]{rWCVPdata::wcvp_names}.
}
\details{
The \code{fuzzy_match} function uses phonetic matching first and then finds
the closest match based on edit distance for any remaining names.

Phonetic matching uses \link[phonics:metaphone]{phonics::metaphone} encoding with a maximum
code length of 20.

Edit distance matching finds the closest match based on Levenshtein similarity,
calculated by \link[RecordLinkage:strcmp]{RecordLinkage::levenshteinSim}.
}
\examples{
wcvp <- rWCVPdata::wcvp_names
fuzzy_match(redlist_example, wcvp, "scientificName")

wcvp <- rWCVPdata::wcvp_names
phonetic_match(redlist_example, wcvp, "scientificName")

wcvp <- rWCVPdata::wcvp_names
edit_match(redlist_example, wcvp, "scientificName")

}
\seealso{
Other name matching functions: 
\code{\link{exact_match}()},
\code{\link{match_names}()}
}
\concept{name matching functions}
