% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/generate_checklist.R
\name{generate_checklist}
\alias{generate_checklist}
\title{Generate a species checklist from WCVP}
\usage{
generate_checklist(
  taxon = NULL,
  taxon.rank = c("species", "genus", "family"),
  area = NULL,
  synonyms = TRUE,
  render.report = FALSE,
  native = TRUE,
  introduced = TRUE,
  extinct = TRUE,
  location_doubtful = TRUE,
  report.filename = NULL,
  local_wcvp = FALSE,
  wcvp_names = NULL,
  wcvp_distributions = NULL
)
}
\arguments{
\item{taxon}{Character. One or many taxa to be included. Defaults to NULL (all species)}

\item{taxon.rank}{Character. One of "species", "genus" or "family" giving the rank of the value/s in \code{taxon}. Must be specified unless taxon is NULL. Note that this must be one value only - it is not possible to mix and match families and genera, for example.}

\item{area}{Character. One or many WGSPRD level 3 region codes. Defaults to NULL (global).}

\item{synonyms}{Logical. Include synonyms in checklist? Defaults to TRUE.}

\item{render.report}{Logical. Render the checklist as a markdown report? Defaults to FALSE.}

\item{native}{Logical. Include species occurrences not flagged as introduced, extinct or doubtful? Defaults to TRUE.}

\item{introduced}{Logical. Include species occurrences flagged as introduced? Defaults to TRUE.}

\item{extinct}{Logical. Include species occurrences flagged as extinct? Defaults to TRUE.}

\item{location_doubtful}{Logical. Include species occurrences flagged as location doubtful? Defaults to TRUE.}

\item{report.filename}{Character. Filename for the HTML file generated. Defaults to requiring user input.}

\item{local_wcvp}{Logical. If FALSE (the default), use data from \code{rWCVPdata}.
If TRUE, use a local copy of the data (useful if rWCVPdata is not the latest
version of the checklist).}

\item{wcvp_names}{Pointer to the WCVP names dataset. Ignored if \code{local.wcvp = FALSE}. Defaults to NULL.}

\item{wcvp_distributions}{Pointer to the WCVP distributions dataset. Ignored if \code{local.wcvp = FALSE}. Defaults to NULL.}
}
\value{

}
\description{
Generate a species checklist from WCVP
}
