% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/generate_occurrence_matrix.r
\name{generate_occurrence_matrix}
\alias{generate_occurrence_matrix}
\title{Generate occurrence matrix for taxa and areas}
\usage{
generate_occurrence_matrix(
  taxon = NULL,
  rank = c("species", "genus", "family", "order", "higher"),
  area = NULL,
  native = TRUE,
  introduced = TRUE,
  extinct = TRUE,
  location_doubtful = TRUE,
  wcvp_names = NULL,
  wcvp_distributions = NULL
)
}
\arguments{
\item{taxon}{Character. One or many taxa to be included. Defaults to NULL (all species)}

\item{rank}{Character. One of "species", "genus", "family", "order" or "higher", giving the rank of the value/s in \code{taxon}. Must be specified unless taxon is \code{NULL}.}

\item{area}{Character. One or many WGSPRD level 3 region codes. Defaults to \code{NULL} (global).}

\item{native}{Logical. Include species occurrences not flagged as introduced, extinct or doubtful? Defaults to TRUE.}

\item{introduced}{Logical. Include species occurrences flagged as introduced? Defaults to TRUE.}

\item{extinct}{Logical. Include species occurrences flagged as extinct? Defaults to TRUE.}

\item{location_doubtful}{Logical. Include species occurrences flagged as location doubtful? Defaults to TRUE.}

\item{wcvp_names}{A data frame of taxonomic names from WCVP version 7 or later.
If \code{NULL}, names will be loaded from \link[rWCVPdata:wcvp_names]{rWCVPdata::wcvp_names}.}

\item{wcvp_distributions}{A data frame of distributions from WCVP version 7 or later.
If \code{NULL}, distributions will be loaded from \link[rWCVPdata:wcvp_names]{rWCVPdata::wcvp_names}.}
}
\value{
A data.frame containing the \code{taxon_name} and \code{plant_name_id}
of all species that are present in the \code{area}, plus one variable for each WGSPRD level 3 region in \code{area}, with species presences marked as 1 and absences marked as 0.
}
\description{
Generate occurrence matrix for taxa and areas
}
\details{
See vignette "Generating occurrence matrices with rWCVP" for an example of how this output can be formatted for publication.
}
\examples{
generate_occurrence_matrix(taxon="Poa",rank="genus",
area=c("TAS", "VIC","NSW"), introduced=FALSE)


}
