% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/wgsrpd3_code_mapping.R
\name{get_wgsrpd3_codes}
\alias{get_wgsrpd3_codes}
\title{Extract WGSRPD Level 3 (area) codes.}
\usage{
get_wgsrpd3_codes(geography, include.equatorial = NULL)
}
\arguments{
\item{geography}{Character. The geography to convert into Level 3 codes. May be a WGSRPD area (Level 3), region (Level 2) or continent (Level 1), country or hemisphere ("Northern Hemisphere", "Southern Hemisphere" or "Equatorial")}

\item{include.equatorial}{Logical. Include Level 3 areas that span the equator? Defaults to Null, which generates a message and includes these areas. Ignored if geography is not a hemisphere.}
}
\value{
Character with area codes (Level 3) that fall within the geography.
}
\description{
Extract WGSRPD Level 3 (area) codes.
}
\examples{
get_wgsrpd3_codes("Brazil")
}
