% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/wcvp_checklist.R
\name{wcvp_checklist}
\alias{wcvp_checklist}
\title{Generate a species checklist from WCVP}
\usage{
wcvp_checklist(
  taxon = NULL,
  taxon.rank = c("species", "genus", "family", "order", "higher"),
  area = NULL,
  synonyms = TRUE,
  render.report = FALSE,
  native = TRUE,
  introduced = TRUE,
  extinct = TRUE,
  location_doubtful = TRUE,
  report.filename = NULL,
  report.dir = NULL,
  report.type = c("alphabetical", "taxonomic"),
  wcvp_names = NULL,
  wcvp_distributions = NULL
)
}
\arguments{
\item{taxon}{Character. Taxon to be included. Defaults to NULL (no taxonomic filter; all taxa).}

\item{taxon.rank}{Character. One of "species", "genus", "family", "order" or "higher", giving the rank of the value/s in \code{taxon}. Must be specified unless taxon is \code{NULL}.}

\item{area}{Character. One or many WGSPRD level 3 region codes. Defaults to \code{NULL} (global).}

\item{synonyms}{Logical. Include synonyms in checklist (see Details)? Defaults to \code{TRUE}.}

\item{render.report}{Logical. Render the checklist as a markdown report? Defaults to FALSE.}

\item{native}{Logical. Include species occurrences not flagged as introduced, extinct or doubtful? Defaults to \code{TRUE}.}

\item{introduced}{Logical. Include species occurrences flagged as introduced? Defaults to \code{TRUE}.}

\item{extinct}{Logical. Include species occurrences flagged as extinct? Defaults to \code{TRUE}.}

\item{location_doubtful}{Logical. Include species occurrences flagged as \code{location_doubtful}? Defaults to \code{TRUE}.}

\item{report.filename}{Character. Name for the HTML file. Defaults to taxon_area_type.html}

\item{report.dir}{Character. Directory for the HTML file to be saved in. Must be provided by user.}

\item{report.type}{Character; one of "alphabetical" or "taxonomic". Should the generated checklist be sorted alphabetically, or by taxonomic status?}

\item{wcvp_names}{A data frame of taxonomic names from WCVP version 7 or later.
If \code{NULL}, names will be loaded from \link[rWCVPdata:wcvp_names]{rWCVPdata::wcvp_names}.}

\item{wcvp_distributions}{A data frame of distributions from WCVP version 7 or later.
If \code{NULL}, distributions will be loaded from \link[rWCVPdata:wcvp_names]{rWCVPdata::wcvp_names}.}
}
\value{
Data.frame with filtered data and, if render.report=TRUE. a report HTML file.
}
\description{
Generate a species checklist from WCVP
}
\details{
The \code{synonyms} argument can be used to limit names to those that are Accepted. If \code{synonyms = TRUE} then invalid, illegitimate and other non-accepted names are also included (i.e., the checklist is not limited to names for which \code{taxon_status == "Synonym"}).
Two styles of checklist are supported in \code{rWCVP} - alphabetical and taxonomic. In an alphabetical checklist, all names are arranged alphabetically with accepted names in bold, and synonyms are followed by their accepted name. For a taxonomic checklist, names are grouped by their accepted names, and synonyms are listed beneath. Both types of checklist include author, publication and distribution information, though note that family headings are only supported in alphabetical checklists (due to the additional grouping requirement of the taxonomic format).
}
\examples{
#These examples take >10 seconds to run.
\dontrun{
wcvp_checklist(taxon="Myrtaceae", taxon.rank="family", area=get_wgsrpd3_codes("Brazil"))
wcvp_checklist(taxon="Ferns", taxon.rank="higher", area=get_wgsrpd3_codes("New Zealand"))
}
}
